# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Nexa REST API
  version: 1.0.0-b01
  summary: Single-node multi-tenant batch AI control plane
  license:
    name: License not declared by B01
    identifier: LicenseRef-Nexa-Unspecified
  description: |
    Contract for browser, CLI, administrator, coordinator-facing worker, and
    local bootstrap operations. Authorization, state transitions, transaction
    semantics, and failure handling are normative in the linked B01 contract
    documents. CLI scopes are exact, non-hierarchical permissions: jobs:read
    covers template/job/session/event/log/result reads; jobs:write covers submit,
    sweep and job controls; artifacts:read and artifacts:write cover tenant
    artifact reads and uploads; tokens:write covers only the caller's token
    metadata lifecycle; admin:read and admin:write cover SYSTEM_ADMIN read and
    mutation operations respectively. Role, tenant membership, ownership, CSRF
    and live authority checks remain mandatory in addition to scope. Unknown
    request fields and duplicate JSON members are rejected.
...
servers:
  - url: https://nexa.example/v1
    description: Deployment-local TLS endpoint; hostname is configuration.
...
paths:
...
  /jobs:
    parameters: [ { $ref: '#/components/parameters/TenantContext' } ]
    get:
      tags: [Jobs]
      operationId: listJobs
      summary: List tenant-owned jobs using keyset pagination
      security: [ { browserCookie: [] }, { cliBearer: [] } ]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
        - name: state
          in: query
          schema: { $ref: '#/components/schemas/JobState' }
        - name: template_id
          in: query
          schema: { type: string, minLength: 1, maxLength: 64 }
        - name: created_after
          in: query
          schema: { $ref: '#/components/schemas/Timestamp' }
      responses:
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
        '200':
          description: Jobs ordered by `created_at DESC, id DESC`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Jobs]
      operationId: submitJob
      summary: Commit one job and logical execution session
      description: Returns 202 only after input/template/capability/admission checks and the job/session/event/idempotency/counter transaction commit.
      security:
        - browserCookie: []
          csrfToken: []
        - cliBearer: []
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JobSubmitRequest' }
            example:
              spec:
                template_id: cpu-iterative
                template_version: 1
                input_artifact_id: 018f0d60-7b6a-7a21-9d82-1aa39c4f30b7
                resources: { cpu_millis: 1000, memory_bytes: 536870912, gpu_count: 0 }
                priority: 1
                runtime_limit_seconds: 120
                checkpoint_interval_seconds: 30
                parameters: { iterations: 100000, seed: 7, modulus: 1000000007 }
      responses:
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
        '202':
          description: Job accepted and durable.
          headers:
            Location: { schema: { type: string, format: uri-reference } }
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Infeasible' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /jobs/{job_id}:
    parameters: [ { $ref: '#/components/parameters/TenantContext' } ]
    get:
      tags: [Jobs]
      operationId: getJob
      summary: Get tenant-owned job state and desired state
      security: [ { browserCookie: [] }, { cliBearer: [] } ]
      parameters: [ { $ref: '#/components/parameters/JobId' } ]
      responses:
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
        '200':
          description: Job representation.
          headers: { ETag: { $ref: '#/components/headers/ETag' } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /sessions/{session_id}:
    parameters: [ { $ref: '#/components/parameters/TenantContext' } ]
    get:
      tags: [Jobs]
      operationId: getLogicalSession
      summary: Get the logical execution session whose state is derived from its job
      security: [ { browserCookie: [] }, { cliBearer: [] } ]
      parameters: [ { $ref: '#/components/parameters/SessionId' } ]
      responses:
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
        '200':
          description: Session identity and derived job state; clients cannot mutate session state.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LogicalSession' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /jobs/{job_id}/events:
    parameters: [ { $ref: '#/components/parameters/TenantContext' } ]
    get:
      tags: [Jobs]
      operationId: listJobEvents
      summary: Read ordered durable events after a sequence
      security: [ { browserCookie: [] }, { cliBearer: [] } ]
      parameters:
        - $ref: '#/components/parameters/JobId'
        - name: after_sequence
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - $ref: '#/components/parameters/PageSize'
      responses:
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
        '200':
          description: Events ordered by `sequence ASC, id ASC`, maximum 100.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /jobs/{job_id}/result:
    parameters: [ { $ref: '#/components/parameters/TenantContext' } ]
    get:
      tags: [Jobs]
      operationId: getJobResult
      summary: Get the single recognized final result
      security: [ { browserCookie: [] }, { cliBearer: [] } ]
      parameters: [ { $ref: '#/components/parameters/JobId' } ]
      responses:
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
        '200':
          description: Fenced unique final result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResultRecord' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /admin/jobs:
    get:
      tags: [Admin]
      operationId: adminListJobs
      summary: Query the cross-tenant queue as a system administrator
      description: Requires an active SYSTEM_ADMIN global grant and admin read scope. Every query is audited; the signed cursor binds the actor and all filters.
      security: [ { browserCookie: [] }, { cliBearer: [] } ]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
        - name: tenant_id
          in: query
          schema: { $ref: '#/components/schemas/UuidV7' }
        - name: user_id
          in: query
          schema: { $ref: '#/components/schemas/UuidV7' }
        - name: state
          in: query
          schema: { $ref: '#/components/schemas/JobState' }
        - name: waiting_reason
          in: query
          schema: { $ref: '#/components/schemas/WaitingReason' }
        - name: created_after
          in: query
          schema: { $ref: '#/components/schemas/Timestamp' }
      responses:
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
        '200':
          description: Jobs ordered by `created_at DESC, id DESC` across authorized tenants.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/jobs/{job_id}:
    get:
      tags: [Admin]
      operationId: adminGetJob
      summary: Read one cross-tenant job as a system administrator
      description: Requires an active SYSTEM_ADMIN global grant and admin read scope; the access is audited and does not grant tenant mutation authority.
      security: [ { browserCookie: [] }, { cliBearer: [] } ]
      parameters: [ { $ref: '#/components/parameters/JobId' } ]
      responses:
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
        '200':
          description: Cross-tenant job representation.
          headers: { ETag: { $ref: '#/components/headers/ETag' } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    browserCookie: { type: apiKey, in: cookie, name: nexa_session }
    csrfToken: { type: apiKey, in: header, name: X-CSRF-Token }
    cliBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-cli-token
      description: Exact non-hierarchical scopes. `jobs:read` authorizes template/job/session/event/log/result reads; `jobs:write` authorizes submit, sweep and job control; `artifacts:read` authorizes tenant artifact list/download; `artifacts:write` authorizes tenant upload; `tokens:write` authorizes only the caller's token list/create/revoke; `admin:read` and `admin:write` authorize SYSTEM_ADMIN reads and mutations. Current-principal introspection accepts any otherwise-valid token. No scope bypasses membership, ownership, global grant, CSRF, tenant context or object-state checks.
    workerBearer: { type: http, scheme: bearer, bearerFormat: opaque-worker-token }
    bootstrapSecret: { type: apiKey, in: header, name: X-Nexa-Bootstrap-Secret }
  headers:
    RequestId: { description: Correlation ID., schema: { $ref: '#/components/schemas/UuidV7' } }
    ETag: { description: Strong version ETag., schema: { type: string, pattern: '^"v[1-9][0-9]*"$' } }
...
    RetryAfter: { description: Minimum whole seconds before retry., schema: { type: integer, minimum: 1 } }
  parameters:
    Cursor: { name: cursor, in: query, required: false, schema: { type: string, minLength: 16, maxLength: 2048 } }
    PageSize: { name: page_size, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
    TenantContext: { name: X-Nexa-Tenant-Id, in: header, required: true, description: Explicit tenant context; active membership/role/scope is rechecked and cursors/idempotency are bound to this value., schema: { $ref: '#/components/schemas/UuidV7' } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, description: 'Replay scope is operation-specific and closed by the contract to tenant UUID, GLOBAL, BOOTSTRAP, WORKER identity, or an Authority-resolved tenant for worker upload; callback operations use X-Callback-Id instead.', schema: { type: string, pattern: '^[!-~]{16,128}$' } }
...
    JobId: { name: job_id, in: path, required: true, schema: { $ref: '#/components/schemas/UuidV7' } }
    SessionId: { name: session_id, in: path, required: true, schema: { $ref: '#/components/schemas/UuidV7' } }
...
  responses:
    BadRequest: { description: 'Malformed body, duplicate JSON member, unknown field, filter, or cursor.', headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } }, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    Unauthorized: { description: 'Missing, invalid, expired, or revoked credential.', headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } }, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    Forbidden: { description: 'Authenticated principal lacks role/scope, tenant membership/context, or CSRF proof.', headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } }, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    NotFound: { description: Resource absent or not visible in the authenticated tenant scope., headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } }, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    Conflict: { description: 'Idempotency key mismatch/in-progress, one-time secret already delivered, callback ID replayed with a different payload, committed state conflict, stale execution authority, or unique conflict. The ErrorResponse code identifies the case. Retry-After is present and equals 1 only for idempotency_in_progress; Location is present only for one_time_secret_unavailable.', headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' }, Retry-After: { $ref: '#/components/headers/RetryAfter' }, Location: { schema: { type: string, format: uri-reference } } }, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
    Infeasible: { description: Request cannot fit total capacity/quota or lacks a configured capability., headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } }, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    RateLimited: { description: Durable user/tenant rate or outstanding-job quota exceeded., headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' }, Retry-After: { $ref: '#/components/headers/RetryAfter' } }, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    DependencyUnavailable: { description: 'Queue full, DB/schema/storage/capability/reconciliation dependency unavailable, or storage pressure.', headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' }, Retry-After: { $ref: '#/components/headers/RetryAfter' } }, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    InternalError: { description: 'Unexpected server failure after safe rollback or fail-closed handling; no credential, path, input, checkpoint, SQL, or cross-tenant data is exposed.', headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } }, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
  schemas:
    UuidV7: { type: string, pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$' }
    Timestamp: { type: string, pattern: '^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}\.[0-9]{3}Z$' }
    Checksum: { type: string, pattern: '^sha256:[0-9a-f]{64}$' }
    Version: { type: integer, format: int64, minimum: 1 }
    ErrorCode: { type: string, enum: [authentication_required, invalid_csrf, permission_denied, resource_not_found, validation_failed, invalid_cursor, idempotency_conflict, idempotency_in_progress, one_time_secret_unavailable, precondition_required, version_conflict, state_conflict, infeasible_request, rate_limited, quota_exceeded, queue_full, dependency_unavailable, checksum_mismatch, payload_too_large, storage_pressure, stale_authority, lease_expired, callback_replayed, internal_error] }
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        message: { type: string, minLength: 1, maxLength: 256 }
        request_id: { $ref: '#/components/schemas/UuidV7' }
    PageInfo:
      type: object
      additionalProperties: false
      required: [next_cursor, page_size]
      properties:
        next_cursor: { type: [string, 'null'], maxLength: 2048 }
        page_size: { type: integer, minimum: 0, maximum: 100 }
...
    ResourceRequestVector:
      type: object
      additionalProperties: false
      required: [cpu_millis, memory_bytes, gpu_count]
      properties:
        cpu_millis: { type: integer, minimum: 100, maximum: 100000000 }
        memory_bytes: { type: integer, format: int64, minimum: 67108864 }
        gpu_count: { type: integer, minimum: 0, maximum: 1 }
...
    CpuParameters:
      type: object
      additionalProperties: false
      required: [iterations, seed, modulus]
      properties:
        iterations: { type: integer, minimum: 1, maximum: 1000000000 }
        seed: { type: integer, minimum: 0, maximum: 2147483647 }
        modulus: { type: integer, minimum: 2, maximum: 2147483647 }
    TrainingParameters:
      type: object
      additionalProperties: false
      required: [epochs, batch_size, learning_rate, seed, subset_size]
      properties:
        epochs: { type: integer, minimum: 1, maximum: 100 }
        batch_size: { type: integer, minimum: 1, maximum: 512 }
        learning_rate: { type: number, exclusiveMinimum: 0, maximum: 1 }
        seed: { type: integer, minimum: 0, maximum: 2147483647 }
        subset_size: { type: integer, minimum: 100, maximum: 50000 }
    InferenceParameters:
      type: object
      additionalProperties: false
      required: [chunk_size, batch_size, output_format]
      properties:
        chunk_size: { type: integer, minimum: 1, maximum: 100000 }
        batch_size: { type: integer, minimum: 1, maximum: 4096 }
        output_format: { type: string, enum: [JSONL, PARQUET] }
    JobSpecBase:
      type: object
      required: [template_id, template_version, input_artifact_id, resources, priority, runtime_limit_seconds, checkpoint_interval_seconds]
      properties:
        template_id: { type: string }
        template_version: { $ref: '#/components/schemas/Version' }
        input_artifact_id: { $ref: '#/components/schemas/UuidV7' }
        resources: { $ref: '#/components/schemas/ResourceRequestVector' }
        priority: { type: integer, minimum: 0, maximum: 2, default: 1 }
        runtime_limit_seconds: { type: integer, minimum: 1, maximum: 300, default: 300 }
        checkpoint_interval_seconds: { type: integer, minimum: 5, maximum: 60, default: 30 }
    CpuJobSpec:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/JobSpecBase'
        - type: object
          required: [template_id, parameters]
          properties:
            template_id: { const: cpu-iterative }
            parameters: { $ref: '#/components/schemas/CpuParameters' }
    TrainingJobSpec:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/JobSpecBase'
        - type: object
          required: [template_id, parameters]
          properties:
            template_id: { const: pytorch-cifar10-cnn }
            parameters: { $ref: '#/components/schemas/TrainingParameters' }
    InferenceJobSpec:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/JobSpecBase'
        - type: object
          required: [template_id, model_artifact_id, parameters]
          properties:
            template_id: { const: batch-inference }
            model_artifact_id: { $ref: '#/components/schemas/UuidV7' }
            parameters: { $ref: '#/components/schemas/InferenceParameters' }
    JobSpec:
      oneOf: [ { $ref: '#/components/schemas/CpuJobSpec' }, { $ref: '#/components/schemas/TrainingJobSpec' }, { $ref: '#/components/schemas/InferenceJobSpec' } ]
      discriminator:
        propertyName: template_id
        mapping:
          cpu-iterative: '#/components/schemas/CpuJobSpec'
          pytorch-cifar10-cnn: '#/components/schemas/TrainingJobSpec'
          batch-inference: '#/components/schemas/InferenceJobSpec'
    JobSubmitRequest:
      type: object
      additionalProperties: false
      required: [spec]
      properties: { spec: { $ref: '#/components/schemas/JobSpec' } }
    JobState: { type: string, enum: [QUEUED, DISPATCHING, RUNNING, PAUSING, PAUSED, RECOVERING, RETRY_WAIT, CANCELLING, SUCCEEDED, FAILED, CANCELLED] }
    DesiredState: { type: string, enum: [RUNNING, PAUSED, CANCELLED] }
    WaitingReason: { type: [string, 'null'], enum: [waiting_for_worker, waiting_for_capacity, waiting_for_quota, waiting_for_reservation, waiting_for_retry, waiting_for_compatibility, null] }
    Job:
      type: object
      additionalProperties: false
      required: [job_id, tenant_id, user_id, session_id, state, desired_state, waiting_reason, spec, spec_checksum, job_fence, retry_count, max_retries, retry_of_job_id, version, event_sequence, created_at, updated_at]
      properties:
        job_id: { $ref: '#/components/schemas/UuidV7' }
        tenant_id: { $ref: '#/components/schemas/UuidV7' }
        user_id: { $ref: '#/components/schemas/UuidV7' }
        session_id: { $ref: '#/components/schemas/UuidV7' }
        state: { $ref: '#/components/schemas/JobState' }
        desired_state: { $ref: '#/components/schemas/DesiredState' }
        waiting_reason: { $ref: '#/components/schemas/WaitingReason' }
        spec: { $ref: '#/components/schemas/JobSpec' }
        spec_checksum: { $ref: '#/components/schemas/Checksum' }
        job_fence: { type: integer, format: int64, minimum: 0 }
        retry_count: { type: integer, minimum: 0, maximum: 2 }
        max_retries: { const: 2 }
        retry_of_job_id: { oneOf: [ { $ref: '#/components/schemas/UuidV7' }, { type: 'null' } ] }
        version: { $ref: '#/components/schemas/Version' }
        event_sequence: { type: integer, format: int64, minimum: 1 }
        created_at: { $ref: '#/components/schemas/Timestamp' }
        updated_at: { $ref: '#/components/schemas/Timestamp' }
    JobPage: { type: object, additionalProperties: false, required: [items, page], properties: { items: { type: array, maxItems: 100, items: { $ref: '#/components/schemas/Job' } }, page: { $ref: '#/components/schemas/PageInfo' } } }
    LogicalSession:
      type: object
      additionalProperties: false
      required: [session_id, job_id, tenant_id, derived_state, created_at]
      properties: { session_id: { $ref: '#/components/schemas/UuidV7' }, job_id: { $ref: '#/components/schemas/UuidV7' }, tenant_id: { $ref: '#/components/schemas/UuidV7' }, derived_state: { $ref: '#/components/schemas/JobState' }, created_at: { $ref: '#/components/schemas/Timestamp' } }
...
    Event:
      type: object
      additionalProperties: false
      required: [event_id, tenant_id, job_id, sequence, type, reason, actor_type, created_at]
      properties: { event_id: { $ref: '#/components/schemas/UuidV7' }, tenant_id: { $ref: '#/components/schemas/UuidV7' }, job_id: { oneOf: [ { $ref: '#/components/schemas/UuidV7' }, { type: 'null' } ] }, sequence: { type: integer, format: int64, minimum: 1 }, type: { type: string, pattern: '^[A-Z][A-Z0-9_]{1,63}$' }, reason: { type: string, minLength: 1, maxLength: 256 }, actor_type: { type: string, enum: [USER, ADMIN, WORKER, COORDINATOR, SYSTEM] }, created_at: { $ref: '#/components/schemas/Timestamp' } }
    EventPage: { type: object, additionalProperties: false, required: [items, page], properties: { items: { type: array, maxItems: 100, items: { $ref: '#/components/schemas/Event' } }, page: { $ref: '#/components/schemas/PageInfo' } } }
...
    ResultRecord:
      type: object
      additionalProperties: false
      required: [result_id, job_id, attempt_id, manifest_artifact_id, manifest_checksum, created_at]
      properties: { result_id: { $ref: '#/components/schemas/UuidV7' }, job_id: { $ref: '#/components/schemas/UuidV7' }, attempt_id: { $ref: '#/components/schemas/UuidV7' }, manifest_artifact_id: { $ref: '#/components/schemas/UuidV7' }, manifest_checksum: { $ref: '#/components/schemas/Checksum' }, created_at: { $ref: '#/components/schemas/Timestamp' } }
...